Move leave room button to sidebar
