+++ src/components/game/LeaveButton.js
@@ -1,0 +1,43 @@
+import React, { useState } from "react";
+import { useTranslation } from "react-i18next";
+import { useHistory, Redirect } from "react-router-dom";
+import { Button, Icon, Confirm } from "semantic-ui-react";
+import { useUser } from "contexts/UserContext";
+import { apiRequests } from "services/API";
+import DataStore from "services/DataStore";
+import { useBoardGame } from "contexts/BoardGameContext";
+import { routes } from "config/routes";
+
+const LeaveButton = () => {
+  const history = useHistory();
+  const { t } = useTranslation("lobby");
+  const { playerID, gameID, gameName, credentials } = useBoardGame();
+  const [confirmOpen, setConfirmOpen] = useState(false);
+  const [error, setError] = useState();
+  const user = useUser();
+
+  const handleLeave = () => {
+    apiRequests
+      .leaveGame(gameName, gameID, playerID, credentials)
+      .then(async () => {
+        await DataStore.deleteCredentials(user.uid, gameID);
+        history.push(routes.lobby());
+      })
+      .catch((e) => {
+        setError(e.message);
+      });
+  };
+
+  return (
+    <>
+      {error && <Redirect pass to={{ pathname: routes.lobby(), state: { error: error } }} />}
+      <Button color="red" onClick={() => setConfirmOpen(true)}>
+        <Icon name="close" />
+        {t("game.leave")}
+      </Button>
+      <Confirm open={confirmOpen} onCancel={() => setConfirmOpen(false)} onConfirm={handleLeave} />
+    </>
+  );
+};
+
+export default LeaveButton;
